Started the wallpaper section + imported wallpapers
I started designing the wallpaper page.
it's going to be very Bleach inspired (if you couldn't tell from the amount of wallpapers I've made).

I added a few wallpapers as well.
I'll have a hero section which shows the latest or best wallpaper, in mostly full size.
The rest of the designs will be featured in a different gallery-like container