Polishing the home page + false page troll
I'm polishing the home page.
Mainly, tinkering with text color, container sizing, removed the top heading with UCT background image and replaced it with a text (AzaamDev) and a logo next to it.

I also started working on an alternative home-page: the persona inspired front-page. (Persona 3 Reload is the main inspiration)